dropdown navigation and gallery edits
this change creates a dropdown navigation menu, which is meant for mobile. as such, with a certain screen size, the menu will collapse and become vertical rather than horizontal.

gallery images were also updated with lower sized images to help with loading. the font sizes were adjusted in the gallery and singular gallery image pages, in order to match the rest of the sites font size changes.